chore(tests): promote one dedupe cluster — consolidate orders CLI tests
Executes the #1083 promote-one workflow:

- Cluster 65e34284f981 (the one whole-file merge in the pending set that
  fits the 400-line hygiene cap): test_commands_orders_editing.py and
  test_commands_orders_history.py fold into test_commands_orders.py
  (370 lines); all three shared the orders_command_test_helpers boundary
  doubles. 15 tests pass unchanged.
- Triage rejections recorded for the two clusters agent-dedupe suggested
  first: 52a37af1e4a7 (7 strategy-engine files, 2283 lines) and
  257be60db787 (6 order-submission files, 1523 lines + conflicting
  autouse fixtures) — whole-file merges there violate the test-hygiene
  cap; the per-concern splits are hygiene-mandated organization.
- Manifest refreshed via agent-dedupe; cluster closed out with owner and
  pr_url per the workflow.

Closes #1083
